requirements.txt
pyyaml
pandas
datasets
huggingface-hub
kaggle
kagglehub